+++ bin/store/src/providers/find/validations.rs
@@ -59,6 +59,12 @@ impl<'a, 'b> Validation<'a, 'b> {
                         return response;
                     }
                 }
+                "concatenated_fields" => {
+                    response = self.validate_concatenated_fields();
+                    if !response.success {
+                        return response;
+                    }
+                }
                 // Add more validation cases as needed
                 _ => {
                     // Handle unknown or unvalidated keys
@@ -123,5 +129,24 @@ impl<'a, 'b> Validation<'a, 'b> {
             count: 0,
             data: vec![],
         }
+    }
+    pub fn validate_concatenated_fields(&self) -> ApiResponse {
+        for concatenate_field in &self.request_body.concatenate_fields {
+            if concatenate_field.fields.is_empty() || concatenate_field.field_name.is_empty() || concatenate_field.entity.is_empty() {
+                return ApiResponse {
+                    success: false,
+                    message: "Each concatenated field must have non-empty fields, field_name and entity".to_string(),
+                    count: 0,
+                    data: vec![],
+                };
+            }
+        }
+
+        return ApiResponse {
+                success: true,
+                message: "Successfully validated concatenated_fields".to_string(),
+                count: 0,
+                data: vec![],
+            };
     }
 }